+++ creme/datasets/gen/__init__.py
@@ -1,0 +1,7 @@
+"""Random data generators."""
+from .sea import SEA
+
+
+__all__ = [
+    'SEA'
+]