+++ CHANGELOG.md
@@ -28,6 +28,8 @@ See progress [here](https://github.com/creme-ml/creme/milestone/1).
 - Added `MaxAbsScaler` and `RobustScaler` to `preprocessing`
 - Added `AbsMax` and `RollingAbsMax` to `stats`
 - Added `PoissonInclusion` to `feature_selection`
+- Added `datasets.gen.SEA`
+- Added `datasets.Higgs`
 
 ### Fixed